backend/requirements.txt
# GroundedAI Backend Dependencies
# All free, open-source packages — Rs.0 total

# --- Core Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# --- Database ---
supabase==2.11.0
httpx==0.28.1

# --- Document Parsing ---
PyMuPDF==1.25.3
python-docx==1.1.2
chardet==5.2.0

# --- Text Processing ---
tiktoken==0.8.0

# --- Embeddings ---
sentence-transformers==3.3.1
torch==2.5.1

# --- Vector Store ---
faiss-cpu==1.9.0.post1

# --- Retrieval ---
rank-bm25==0.2.2
numpy==1.26.4

# --- LLM (Ollama client via httpx, already included) ---

# --- Caching ---
redis==5.2.1

# --- Evaluation ---
rouge-score==0.1.2
nltk==3.9.1
reportlab==4.2.5

# --- Logging ---
structlog==24.4.0

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.24.0

# --- Code Quality ---
black==24.10.0
flake8==7.1.1
isort==5.13.2

# --- Upload handling ---
python-multipart==0.0.20
